fix(#2163): verify pruned candidates under switch; serialize global-state tests (roborev r5)
